requirements.txt
astroid==2.1.0
certifi==2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==2.1
django-braces==1.13.0
django-cors-headers==2.4.0
django-cors-middleware==1.3.1
django-environ==0.4.5
django-extensions==2.1.4
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.1.0
django-taggit==0.23.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.7
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
oauthlib==2.1.0
openapi-codec==1.3.2
psycopg2-binary==2.7.6.1
PyJWT==1.6.4
pylint==2.2.1
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.7
requests==2.20.1
requests-oauthlib==1.0.0
simplejson==3.16.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1.2
wrapt==1.10.11
psycopg2==2.7.6.1
